requirements.txt
Brotli==1.0.9
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
dash==1.19.0
dash-core-components==1.15.0
dash-html-components==1.1.2
dash-renderer==1.9.0
dash-table==4.11.2
Flask==1.1.2
Flask-Compress==1.9.0
future==0.18.2
google-api-core==1.26.1
google-auth==1.27.1
google-cloud-core==1.6.0
google-cloud-logging==2.2.0
googleapis-common-protos==1.53.0
grpcio==1.36.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.20.1
packaging==20.9
pandas==1.2.3
plotly==4.14.3
proto-plus==1.17.0
protobuf==3.15.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
retrying==1.3.3
rsa==4.7.2
six==1.15.0
urllib3==1.26.3
Werkzeug==1.0.1
gunicorn